fixed skill percentage int to float

+++ src/main/java/com/thomaspufahl/apiportfolio/Portfolio/Skill/Skill.java
@@ -18,7 +18,7 @@ public class Skill {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
-    private int percentage;
+    private float percentage;
 
     @ManyToOne
     private Person person;
